requirements.txt
rich==13.7.0
tabulate==0.9.0
python-dateutil==2.8.2
